work on game resource types

+++ branches/v2/library/Zupal/Domain/Abstract.php
@@ -7,10 +7,10 @@
  */
 abstract class Zupal_Domain_Abstract
 implements Zupal_Domain_IDomain {
-
+    
     const STUB = '_asStub_';
     protected $_row = NULL;
-
+    
     /**
      * Creates an object based on the passed ID.
      * NOTE: there is NO redundancy protection on domain objects because
@@ -27,9 +27,11 @@ implements Zupal_Domain_IDomain {
             $this->load($pID);
         else:
             $this->newRow();
-    endif;
+        endif;
+        
+        $this->init();
     }
-
+    
     public static function _as($pItem, $pClass, $pAsID = FALSE) {
         if (!$pItem instanceof $pClass):
             if (is_scalar($pItem)):
@@ -41,16 +43,23 @@ implements Zupal_Domain_IDomain {
                 throw new Exception(__METHOD__ . ': cannot convert ' . print_r($pItem, 1) . ' to ' . $pClass);
         endif;
         endif;
-
+        
         if ($pAsID):
             return $pItem->identity();
         else:
             return $pItem;
     endif;
     }
-
+    
+/* @@@@@@@@@@@@@@@@@@@@@@@@@@@@@ init @@@@@@@@@@@@@@@@@@@@@@@@@@@@@ */
+    /**
+     * an extension point for new records
+     */
+    public function init () {
+    }
+    
 /* @@@@@@@@@@@@@@@@@@@@@@@@@@@@@ new @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@ */
-
+    
     protected function newRow($pData = NULL) {
         $this->_row = $this->table()->fetchNew();
         $row = $this->_row;
@@ -68,22 +77,22 @@ implements Zupal_Domain_IDomain {
                 endforeach;
         endif;
         endif;
-
+        
         return $row;
     }
-
+    
     protected $_joins = array();
-
+    
     public function get_join($pKey, $pCreate_if_empty = TRUE) {
     //	$pType = strtolower(trim($pType));
-
+        
         $data = $this->_joins[$pKey];
         if ($data):
             extract($data);
             $id = $this->__get($local_key);
-
+            
             extract($data); // should return $class, $local_key, $value = NULL
-
+            
             if ($id):
                 if ($value):
                     if ($value->identity() != $id):
@@ -103,7 +112,7 @@ implements Zupal_Domain_IDomain {
         endif;
         return NULL;
     }
-
+    
     public function set_join($pKey, $pClass, $pID = NULL, $pLocal_Key = NULL) {
         if ($pLocal_Key):
             $data = array(
@@ -120,7 +129,7 @@ implements Zupal_Domain_IDomain {
             return NULL;
     endif;
     }
-
+    
 /* @@@@@@@@@@@@@@@@@@@@@@@@@@@@@ table @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@ */
     /**
      * @return Zupal_Table_Abstract
@@ -130,16 +139,16 @@ implements Zupal_Domain_IDomain {
         if (! array_key_exists($tc, self::$_tables)) {
             self::$_tables[$tc] = new $tc();
         }
-
+        
         return self::$_tables[$tc];
     }
-
+    
     private static $_tables = array();
-
+    
     protected abstract function tableClass ();
-
+    
 /* @@@@@@@@@@@@@@@@@@@@@@@@@@@@@ identity @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@ */
-
+    
     public function identity() {
         if ($this->isStub()) return NULL;
         //@NOTE: elaborate retrieval for debugging purposes -- should be simplified for production.
@@ -156,12 +165,12 @@ implements Zupal_Domain_IDomain {
 			throw new Exception(__METHOD__ . ': non object row in ' . $this->tableClass() . ': ' . print_r($this->_row, 1));
 		endif; */
     }
-
+    
 /* @@@@@@@@@@@@@@@@@@@@@@@@@@@@@ load @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@ */
 /*
  * Loads the row with data from the database -- or mock source.
  */
-
+    
     protected function load ($pID) {
         if (is_object($pID)):
             if ($pID instanceof Zend_Db_Table_Row):
@@ -171,7 +180,7 @@ implements Zupal_Domain_IDomain {
                 throw new Exception(__METHOD__ . ': Non integer ' . print_r($pID, 1) . ' passed to ' . __CLASS__);
         endif;
         endif;
-
+        
         if ($pID):
             $hits = $this->table()->find($pID);
             if ($hits):
@@ -183,9 +192,9 @@ implements Zupal_Domain_IDomain {
         endif;
         if (!$this->_row) $this->_row = $this->table()->createRow();
     }
-
+    
 /* @@@@@@@@@@@@@@@@@@@@@@@@@@@@@ Status tests @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@ */
-
+    
     /**
      * isSaved and loaded answer two similar questions - but the difference is important.
      * isSaved uses the database as the system of record. If the loaded ID is in the database (or is zero)
@@ -199,11 +208,11 @@ implements Zupal_Domain_IDomain {
         $id_field = $this->table()->idField();
         $table_name = $this->table()->tableName();
         $id = $this->identity();
-
+        
         if (!$id):
-
+            
             return FALSE;
-
+            
         else:
             if (is_numeric($id)):
                 $sql = "SELECT count(`$id_field`) FROM `$table_name` WHERE `$id_field` = $id";
@@ -211,12 +220,12 @@ implements Zupal_Domain_IDomain {
                 $sql = "SELECT count(`$id_field`) FROM `$table_name` WHERE `$id_field` LIKE '$id'";
             endif;
             $tally = $this->table()->getAdapter()->fetchOne($sql);
-
+            
             return $tally; // note -- any (unlikely) duplication of an ID key in a table has to be handled downstream of this method
-
+    
     endif;
     }
-
+    
     public function loaded() {
         if (!$this->_row) return FALSE;
         if (is_object($this->_row)):
@@ -227,9 +236,9 @@ implements Zupal_Domain_IDomain {
             return FALSE;
     endif;
     }
-
+    
 /* @@@@@@@@@@@@@@@@@@@@@@@@@@@@@ find @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@ */
-
+    
     /**
      * return one (or more) row; should at least be able to find a single
      * domain object; more complex queries should be managed by the domain
@@ -244,7 +253,7 @@ implements Zupal_Domain_IDomain {
      public abstract function find ($pParams, $pSort = NULL);
      */
 /* @@@@@@@@@@@@@@@@@@@@@@@@@@@@@ find @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@ */
-
+    
     /**
      * Your basic query gateway.
      * Find returns an array or rowset of
@@ -253,14 +262,14 @@ implements Zupal_Domain_IDomain {
      *
      * NOTE: cannot handle joins -- use find_from_sql with table = false for join based results.
      */
-
+    
     public function find($pParams = NULL, $pSort = NULL) {
         $rows = array();
         if (is_numeric($pParams)):
             $rows = $this->get($pParams);
         elseif (is_array($pParams)):
             $select = $this->_select($pParams, $pSort);
-
+            
             $table_rows = $this->table()->fetchAll($select);
         elseif (is_null($pParams) || ($pParams instanceof Zend_Db_Table_Select)):
             $table_rows = $this->table()->fetchAll($pParams, $pSort);
@@ -270,18 +279,18 @@ implements Zupal_Domain_IDomain {
                 $rows[] = $this->get($row);
             endforeach;
         endif;
-
+        
         return $rows;
     }
-
+    
 /* @@@@@@@@@@@@@@@@@@@@@@@@@@@@@ findAll @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@ */
-
+    
     public function findAll ($pSort = NULL) {
         return $this->find(NULL, $pSort);
     }
-
+    
 /* @@@@@@@@@@@@@@@@@@@@@@@@@@@@@ findOne @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@ */
-
+    
     /**
      * @return Zupal_Domain_Abstract.
      * If no match is found, NULL returns. 
@@ -291,7 +300,7 @@ implements Zupal_Domain_IDomain {
         $row = $this->table()->fetchRow($select);
         return $row ? $this->get($row) : NULL;
     }
-
+    
 /* @@@@@@@@@@@@@@@@@@@@@@@@@@@@@ _select @@@@@@@@@@@@@@@@@@@@@@@@@@@@@ */
     /**
      *
@@ -316,18 +325,18 @@ implements Zupal_Domain_IDomain {
                     $select->where("`$field` $operator $value");
                 else:
                     $select->where("`$field` $operator ?", $value);
-                endif;
+            endif;
             endforeach;
         endif;
         if ($pSort):
             $select->order($pSort);
         endif;
-      //  $sql = $select->assemble();
+        //  $sql = $select->assemble();
         return $select;
     }
-
+    
 /* @@@@@@@@@@@@@@@@@@@@@@@@@@@@@ get @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@ */
-
+    
     /**
      * returns a domain object for the passed ID.
      * NOTE: does not check for the EXISTENCE of a row with said ID.
@@ -339,20 +348,20 @@ implements Zupal_Domain_IDomain {
      * @return Zupal_Domain_Abstract
      */
     public abstract function get ($pID = NULL, $pField_Values = NULL);
-
+    
 /* @@@@@@@@@@@@@@@@@@@@@@@@@@@@@ __get @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@ */
-
+    
     public function __get ($pField) {
         return $this->_row->__get($pField);
     }
-
+    
 /* @@@@@@@@@@@@@@@@@@@@@@@@@@@@@ set @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@ */
-
+    
     public function __set($pField, $pValue) {
     //@NOTE: __get accepts array data but this method presumes row object. One or the other direction needs to be solid.
         $this->_row->__set($pField, $pValue);
     }
-
+    
 /* @@@@@@@@@@@@@@@@@@@@@@@@@@@@@ set_fields @@@@@@@@@@@@@@@@@@@@@@@@@@@@@ */
     /**
      *
@@ -369,19 +378,19 @@ implements Zupal_Domain_IDomain {
                 catch (Exception $e) {
                     error_log(__METHOD__ . ": bad field assignamtion: $f => $v");
                 }
-            endif;
-
+        endif;
+        
         endforeach;
     }
-
+    
 /* @@@@@@@@@@@@@@@@@@@@@@@@@@@@@ save @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@ */
-
+    
     public function save() {
     //@NOTE: __get accepts array data but this method presumes row object. One or the other direction needs to be solid.
         if ($this->_asStub):
             throw new Exception('Attempt to save a stub of ' . get_class($this));
         elseif ($this->_row):
-
+            
             foreach($this->_joins as $join):
                 if ($join && is_array($join)):
                     $value = NULL;
@@ -390,35 +399,49 @@ implements Zupal_Domain_IDomain {
                         $value->save();
                         $this->$local_key = $value->identity();
                 endif;
-
+                
             endif;
             endforeach;
             if ($this->_row instanceof stdClass):
                 print_r($this->_row);
                 throw new Exception('Cannot save ' . print_r($this->_row, 1));
             endif;
-
+            
             $this->_row->save();
-
+            
         else:
             throw new Exception(__METHOD__ . ': Cannot save empty row');
     endif;
     }
-
+    
 /* @@@@@@@@@@@@@@@@@@@@@@@@@@@@@ asStub @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@ */
-
+    
     protected $_asStub = FALSE;
     public function asStub() {
         $this->_asStub = TRUE;
     }
     public function isStub() { return $this->_asStub; }
-
+    
 /* @@@@@@@@@@@@@@@@@@@@@@@@@@@@@ delete @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@ */
-
-    public function delete() {
-        $this->_row->delete();
+    
+    protected $_soft_delete = FALSE;
+    protected $_soft_delete_key = 'active';
+    
+    public function delete($pTotal = FALSE) {
+        if ($this->_soft_delete):
+            if ($pTotal):
+                $this->_row->delete();
+            else:
+                $sdk = $this->_soft_delete_key;
+                $this->$sdk = 0;
+                $this->save();
+                return;
+        endif;
+        else:
+            $this->_row->delete();
+    endif;
     }
-
+    
 /* @@@@@@@@@@@@@@@@@@@@@@@@@@@@@ toArray @@@@@@@@@@@@@@@@@@@@@@@@@@@@@ */
     /**
      *
@@ -427,9 +450,9 @@ implements Zupal_Domain_IDomain {
     public function toArray () {
         return $this->_row->toArray();
     }
-
+    
 /* @@@@@@@@@@@@@@@@@@@@@@@@@@@@@ find_from_sql @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@ */
-
+    
     /**
      * This method takes on faith that the SQL has been constructed
      * to return an array of (at least) identities,
@@ -438,22 +461,22 @@ implements Zupal_Domain_IDomain {
      * @TODO: validate this!
      * @return Zupal_Domain_Abstract[];
      */
-
+    
     public function find_from_sql($pSQL, $pTable = TRUE, $pBy_ID = TRUE) {
         if ($pTable):
             $base = $this->table();
         else:
             $base = $this->table()->getAdapter();
         endif;
-
+        
         if (is_array($pSQL)):
             $rowset = call_user_func_array(array($base, 'fetchAll'), $pSQL);
         else:
             $rowset = $base->fetchAll($pSQL);
         endif;
-
+        
         $rows = array();
-
+        
         if ($pBy_ID):
             $id_field = $this->table()->idField();
             foreach ($rowset as $data):
@@ -464,14 +487,14 @@ implements Zupal_Domain_IDomain {
                 $rows[] = $this->get($data);
             endforeach;
         endif;
-
+        
         return $rows;
     }
-
+    
 	/* @@@@@@@@@@@@@@@@@@@@@@@@@@@@@ link @@@@@@@@@@@@@@@@@@@@@@@@@@@@@ */
-
+    
     const LINK_TEMPLATE = '<a href="%s" %s>%s</a>';
-
+    
     /**
      * returns a markup link to this item
      *
@@ -481,10 +504,10 @@ implements Zupal_Domain_IDomain {
      */
     public function link ($pURL, $pClass = NULL) {
         $class = $pClass ? sprintf(' class="%s" ', $pClass) : '';
-
+        
         return sprintf(self::LINK_TEMPLATE, $pURL, $class, $this);
     }
-
+    
 /* @@@@@@@@@@@@@@@@@@@@@@@@@@@@@ __call @@@@@@@@@@@@@@@@@@@@@@@@@@@@@ */
     /**
      *
@@ -502,10 +525,10 @@ implements Zupal_Domain_IDomain {
         if (method_exists($this, $alt)):
             return call_user_func_array(array($this, $alt), $pParams);
         endif;
-
+        
         throw new Exception("No function $pName or $alt in " . get_class($this));
     }
-
+    
 /* @@@@@@@@@@@@@@@@@@@@@@@@@@@@@ enum_field @@@@@@@@@@@@@@@@@@@@@@@@@@@@@ */
     /**
      *
@@ -513,19 +536,111 @@ implements Zupal_Domain_IDomain {
      * @return <type>
      */
     private $_enum_fields = NULL;
-
+    
     public function enum_field ($pField) {
         if (!is_array($this->_enum_fields)):
             $this->_enum_fields = array();
         endif;
-
+        
         if (!array_key_exists($pField, $this->_enum_fields)):
             $this->_enum_fields[$pField] = new Zupal_Domain_Enum($this, $pField);
         endif;
-
+        
         return $this->_enum_fields[$pField];
     }
+    
+/* @@@@@@@@@@@@@@@@@@@@@@@@@@@@@ move @@@@@@@@@@@@@@@@@@@@@@@@@@@@@ */
+    /**
+     *
+     * @param Zupal_Domain_Abstract $pItem
+     */
+    public function move ($pItem, $pMode, $pRank, $pData = NULL) {
+        
+        $id = $pItem->identity();
+        
+        if (!$pItem->isSaved()):
+            throw new Exception(__METHOD__ . ': attempt to move unsaved ' . get_class($pItem));
+        endif;
 
+        if (is_null($pData)):
+            $pData = $this->findAll($pRank);
+        endif;
+        
+                $found = FALSE;
+                $out = array();
+                
+        switch($pMode):
+            case self::MOVE_TOP:
+                $out [] = $pItem;
+                
+                foreach($pData as $item):
+                    if ($pItem->identity() != $item->identity()):
+                        $out[] = $item;
+                endif;
+                endforeach;
+                break;
+                
+            case self::MOVE_UP:
+                
+                foreach($pData as $item):
+                    if ($found):
+                        $out[] = $item;                    
+                    elseif ($pItem->identity() == $item->identity()):
+                        if (count($out)):
+                            $prev = array_pop($out);
+                            $out[] = $item;
+                            $out[] = $prev;
+                        else:
+                            $out[] = $item;
+                        endif;
+                        $found = TRUE;
+                    else:
+                        $out[] = $item;
+                    endif;
+                endforeach;
+                break;
+                
+            case self::MOVE_DOWN:
+                $just_saved = FALSE;
+                foreach($pData as $item):
+                    if ($found):
+                        $out[] = $item;  
+                        if ($just_saved):
+                            $out[] = $pItem;
+                            $just_saved = FALSE;
+                        endif;
+                    elseif ($pItem->identity() == $item->identity()):
+                        $found = TRUE;
+                        $just_saved = TRUE;
+                    endif;
+                endforeach;
+                    
+                break;
+            case self::MOVE_BOTTOM:
+                
+                foreach($pData as $item):
+                    if ($pItem->identity() != $item->identity()):
+                        $out[] = $item;
+                endif;
+                endforeach;
+                $out [] = $pItem;
+                break;
+
+            default:
+                throw new Exception(__METHOD__  . ': attempt to sort with unknown mode ' . $pMode);
+        endswitch;
+
+       foreach($out as $k => $item):
+           $item->$pRank = $k;
+           $item->save();
+       endforeach;
+
+    }
+    
+    const MOVE_UP = 'move_up';
+    const MOVE_TOP = 'move_top';
+    const MOVE_DOWN = 'move_down';
+    const MOVE_BOTTOM = 'move_bottom';
 
 }
